+++ LinxRun/LinxRun/Cosmics.h
@@ -92,7 +92,7 @@ Raster<typename TPsf::Value> quotient(const TIn& in, const TPsf& psf)
   auto filter = StructuringElement<QuotientFilter<T>, Box<2>>(
       QuotientFilter<T>(psf.begin(), psf.end()),
       psf.domain() - (psf.shape() - 1) / 2);
-  return filter * extrapolate<NearestNeighbor>(in);
+  return filter * extrapolation<NearestNeighbor>(in);
 }
 
 template <typename TIn, typename TPsf>
@@ -102,7 +102,7 @@ Raster<typename TPsf::Value> match(const TIn& in, const TPsf& psf)
   auto filter = StructuringElement<PearsonCorrelation<T>, Box<2>>(
       PearsonCorrelation<T>(psf.begin(), psf.end()),
       psf.domain() - (psf.shape() - 1) / 2);
-  return filter * extrapolate<NearestNeighbor>(in);
+  return filter * extrapolation<NearestNeighbor>(in);
 }
 
 template <typename TIn>
@@ -111,7 +111,7 @@ Raster<typename TIn::Value> laplacian(const TIn& in)
   using T = typename TIn::Value;
   const auto filter = convolution(
       Raster<T>({3, 3}, {-1. / 6., -2. / 3., -1. / 6., -2. / 3., 10. / 3., -2. / 3., -1. / 6., -2. / 3., -1. / 6.}));
-  return filter * extrapolate<NearestNeighbor>(in);
+  return filter * extrapolation<NearestNeighbor>(in);
 }
 
 template <typename TIn>
@@ -119,7 +119,7 @@ Raster<typename TIn::Value> dilate(const TIn& in, Index radius = 1)
 {
   using T = typename TIn::Value;
   auto filter = dilation<T>(Box<2>::from_center(radius)); // FIXME L2-ball
-  return filter * extrapolate<NearestNeighbor>(in);
+  return filter * extrapolation<NearestNeighbor>(in);
 }
 
 template <typename TIn>
@@ -127,7 +127,7 @@ Raster<typename TIn::Value> blur(const TIn& in, Index radius = 1)
 {
   using T = typename TIn::Value;
   auto filter = mean_filter<T>(Box<2>::from_center(radius)); // FIXME L2-ball
-  return filter * extrapolate<NearestNeighbor>(in);
+  return filter * extrapolation<NearestNeighbor>(in);
 }
 
 /**
@@ -210,7 +210,7 @@ template <typename TIn, typename TMask>
 void segment(const TIn& in, TMask& mask, float threshold)
 {
   // FIXME Mask<2>::ball<1>(1)
-  auto candidates = dilation<typename TMask::Value>(Box<2>::from_center(1)) * extrapolate(mask, '\0') - mask;
+  auto candidates = dilation<typename TMask::Value>(Box<2>::from_center(1)) * extrapolation(mask, '\0') - mask;
   for (const auto& p : candidates.domain() - Box<2>::from_center(1)) {
     if (candidates[p]) {
       if (min_contrast(in, mask, p) < threshold) {