+++ LinxTransforms/tests/src/StructuringElement_test.cpp
@@ -19,7 +19,7 @@ BOOST_AUTO_TEST_CASE(constant0_3x3_test)
 {
   Raster<int, 2> in({4, 3});
   in.fill(1);
-  const auto extra = extrapolate(in, 0);
+  const auto extra = extrapolation(in, 0);
   const auto box = Box<2>::from_center(1);
 
   auto median_out = median_filter<int>(box) * extra;